feat: implement stream stats with file list (#1433)

+++ src/job/metrics.rs
@@ -205,10 +205,10 @@ async fn update_storage_metrics() -> Result<(), anyhow::Error> {
             .set(stat.compressed_size as i64);
         metrics::STORAGE_FILES
             .with_label_values(&[columns[0], columns[2], columns[1]])
-            .set(stat.file_num as i64);
+            .set(stat.file_num);
         metrics::STORAGE_RECORDS
             .with_label_values(&[columns[0], columns[2], columns[1]])
-            .set(stat.doc_num as i64);
+            .set(stat.doc_num);
     }
     Ok(())
 }
